# test-harness/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "APM Checkout SDK — Test Harness API",
    "version": "1.0.0",
    "description": "Backend API for the APM Checkout SDK test harness. Proxies API calls to sandbox payment providers (Klarna, CashApp, PPRO) and provides mock endpoints for other APMs."
  },
  "servers": [
    { "url": "http://localhost:3847", "description": "Local test harness" }
  ],
  "paths": {
    "/api/klarna/session": {
      "post": {
        "summary": "Create Klarna payment session",
        "tags": ["Klarna"],
        "description": "Creates a Klarna Payments session via the playground sandbox. Returns a client_token for widget initialization.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": { "type": "number", "example": 49.99 },
                  "currency": { "type": "string", "example": "USD" },
                  "taxAmount": { "type": "number", "example": 0 },
                  "merchantReference": { "type": "string", "example": "SDK-1234567890" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "itemName": { "type": "string" },
                        "quantity": { "type": "integer" },
                        "unitPrice": { "type": "number" },
                        "grossAmount": { "type": "number" },
                        "taxAmount": { "type": "number" }
                      }
                    }
                  },
                  "shippingAddress": { "type": "object" },
                  "billingAddress": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "paymentMethod": { "type": "object" },
                    "order": { "type": "object" },
                    "_raw": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}